+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 * Added CreatePopupFromResources, CreatePopupFromText, and CreatePopupFromObject functions
 * Added implementation of Properties interface to Popup
 * Changed ParseDataText function return values
+* Added `Properties() iter.Seq[DataNode]` iterator to DataObject interface
+* Renamed `ArrayElements() []DataValue` method of DataNode interface to `Array() []DataValue` 
+* Added `ArrayElements() iter.Seq[DataValue]` iterator to DataNode interface
+
 
 # v0.19.0